Fix CodeFactor Issues

+++ src/Light/LightManager.cpp
@@ -14,50 +14,37 @@ void LightManager::Debug() {
             if (ImGui::BeginTabBar("Light")){
                 if (ImGui::BeginTabItem("General")){
                     if(ImGui::CollapsingHeader("Files")){
-                        if(ImGui::Button("Load / Reload")){Load();}
-                        ImGui::SameLine();
-                        if(ImGui::Button("Save")){Save();}
+                        if(ImGui::Button("Load / Reload")){Load();} ImGui::SameLine(); if(ImGui::Button("Save")){Save();}
                     }
-
                     if (ImGui::CollapsingHeader("Count")){
-                        ImGui::Text("Directional: %d", lightCount_->dlCount);
-                        ImGui::Text("Point: %d", lightCount_->plCount);
-                        ImGui::Text("Spot: %d", lightCount_->slCount);
+                        ImGui::Text("Directional: %d", static_cast<int>(lightCount_->dlCount));
+                        ImGui::Text("Point: %d", static_cast<int>(lightCount_->plCount));
+                        ImGui::Text("Spot: %d", static_cast<int>(lightCount_->slCount));
                     }
-
                     ImGui::PushID("Directional");
                     ImGui::SeparatorText("Directional");
                     if (ImGui::Button("Add")){Add(LightType::Directional);}
                     ImGui::PopID();
-
                     ImGui::PushID("Point");
                     ImGui::SeparatorText("Point");
                     if (ImGui::Button("Add")){Add(LightType::Point);}
                     ImGui::PopID();
-
                     ImGui::PushID("Spot");
                     ImGui::SeparatorText("Spot");
                     if (ImGui::Button("Add")){Add(LightType::Spot);}
                     ImGui::PopID();
-
                     ImGui::EndTabItem();
                 }
                 if (ImGui::BeginTabItem("Directional")){
-                    for (const auto& dl : rawDirectionalLights_){
-                        dl->Update();
-                    }
+                    for (const auto& dl : rawDirectionalLights_){dl->Update();}
                     ImGui::EndTabItem();
                 }
                 if (ImGui::BeginTabItem("Point")){
-                    for (const auto& pl : rawPointLights_){
-                        pl->Update();
-                    }
+                    for (const auto& pl : rawPointLights_){pl->Update();}
                     ImGui::EndTabItem();
                 }
                 if (ImGui::BeginTabItem("Spot")){
-                    for (const auto& sl : rawSpotLights_){
-                        sl->Update();
-                    }
+                    for (const auto& sl : rawSpotLights_){sl->Update();}
                     ImGui::EndTabItem();
                 }
                 ImGui::EndTabBar();
@@ -145,7 +132,6 @@ void LightManager::Load() {
 }
 
 void LightManager::Save() const {
-
     for (const auto& dl : rawDirectionalLights_){
         dl->Save(path);
     }
